fix: make config_c9io_machine.sh reentrant

+++ config_c9io_machine.sh
@@ -8,10 +8,12 @@ if [ -z "${C9_USER}" ]; then
   exit 1
 fi
 
+CURRENT_TIMESTAMP="$(date +"%Y%m%d_%H%M%S")"
+
 set -eox pipefail
 
 # Create a log file of the script output
-LOG_FILENAME="$(basename $0 .sh)_$(date +"%Y%m%d_%H%M%S").log"
+LOG_FILENAME="$(basename $0 .sh)_${CURRENT_TIMESTAMP}.log"
 exec >  >(tee -a ${LOG_FILENAME})
 exec 2> >(tee -a ${LOG_FILENAME} >&2)
 
@@ -19,34 +21,36 @@ exec 2> >(tee -a ${LOG_FILENAME} >&2)
 sudo mv /etc/localtime /etc/localtime.bak && sudo ln -s /usr/share/zoneinfo/Europe/Zurich /etc/localtime && date
 
 # tmux
-[ -f ~/.tmux.conf ] && cp -v ~/.tmux.conf ~/.tmux.conf.backup_$(date +"%Y%m%d%H%M%S")
+[ -f ~/.tmux.conf ] && cp -v ~/.tmux.conf ~/.tmux.conf.orig_${CURRENT_TIMESTAMP}
 curl -L https://gist.githubusercontent.com/rbf/3529029/raw/.tmux.conf -o ~/.tmux.conf
 # Disable tmux option not valid on Cloud9. Might be related to https://github.com/tony/tmux-config/issues/26
-sed -i.orig -e 's/set -g status-utf8 on/# set -g status-utf8 on/' ~/.tmux.conf
+sed -i.orig_${CURRENT_TIMESTAMP} -e 's/set -g status-utf8 on/# set -g status-utf8 on/' ~/.tmux.conf
 # SOURCE: https://sanctum.geek.nz/arabesque/reloading-tmux-config/
+/mnt/shared/sbin/tmux source-file ~/.tmux.conf
 # NOTE: tmux path found with 'ps -u | grep tmux'
 # SOURCE: http://stackoverflow.com/a/26705778
-/mnt/shared/sbin/tmux source-file ~/.tmux.conf
 
 # gitconfig
-[ -f ~/.gitconfig ] && cp -v ~/.gitconfig ~/.gitconfig.backup_$(date +"%Y%m%d%H%M%S")
+[ -f ~/.gitconfig ] && cp -v ~/.gitconfig ~/.gitconfig.orig_${CURRENT_TIMESTAMP}
 curl -L https://gist.githubusercontent.com/raw/1845578/.gitconfig -o ~/.gitconfig
 
 # git ignore
-if [ -f ~/.gitignore_global ]; then cp -v ~/.gitignore_global ~/.gitignore_global.backup; fi;
+if [ -f ~/.gitignore_global ]; then cp -v ~/.gitignore_global ~/.gitignore_global.orig_${CURRENT_TIMESTAMP}; fi;
 curl -L https://gist.githubusercontent.com/rbf/2224744/raw/.gitignore_global -o ~/.gitignore_global
 git config --global core.excludesfile ~/.gitignore_global
 
-# Update PostgreSQL to 9.4 (default version on Cloud9 as of this writing is 9.3.13)
-# SOURCE: https://community.c9.io/t/can-we-upgrade-to-postgres-9-4/3897/4
-psql --version
-sudo service postgresql stop
-sudo apt-get -y --purge remove postgresql\*
-sudo sh -c 'echo "deb http://apt.postgresql.org/pub/repos/apt/ trusty-pgdg main" >> /etc/apt/sources.list.d/pgdg.list'
-wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
-sudo apt-get -y update
-sudo apt-get -y install postgresql-9.4
-psql --version
+if [ -z "$(psql --version | grep 9.4)" ]; then
+  # Update PostgreSQL to 9.4 (default version on Cloud9 as of this writing is 9.3.13)
+  # SOURCE: https://community.c9.io/t/can-we-upgrade-to-postgres-9-4/3897/4
+  psql --version
+  sudo service postgresql stop
+  sudo apt-get -y --purge remove postgresql\*
+  sudo sh -c 'echo "deb http://apt.postgresql.org/pub/repos/apt/ trusty-pgdg main" >> /etc/apt/sources.list.d/pgdg.list'
+  wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
+  sudo apt-get -y update
+  sudo apt-get -y install postgresql-9.4
+  psql --version
+fi
 
 # Create a DB with UTF8 encoding.
 
@@ -56,14 +60,19 @@ sudo service postgresql stop
 sudo pg_dropcluster 9.4 main
 # SOURCE: http://dba.stackexchange.com/a/19585
 sudo pg_createcluster 9.4 main -e UTF8
+
 # Allow DB users to connect without password.
 # SOURCE: https://community.c9.io/t/can-we-upgrade-to-postgres-9-4/3897/4
 # SOURCE: http://stackoverflow.com/a/9630739
-sudo sed -i.orig -e 's/peer$/trust/' -e 's/md5$/trust/' /etc/postgresql/9.4/main/pg_hba.conf
+sudo sed -i.orig_${CURRENT_TIMESTAMP} -e 's/peer$/trust/' -e 's/md5$/trust/' /etc/postgresql/9.4/main/pg_hba.conf
 sudo service postgresql start
-# Create the DB user defined in /config/database.yml
+
+# Create a DB superuser for the current ubuntu user to be able to use launch DB commands without 'sudo su - postgres ...'
+# SOURCE: http://stackoverflow.com/a/8546783/5764181
 # SOURCE: http://dba.stackexchange.com/a/19573
-sudo su - postgres -c 'createuser --superuser basimilchdbuser'
+sudo su - postgres -c "psql postgres -tAc \"SELECT 1 FROM pg_roles WHERE rolname='$(whoami)'\" | grep -q 1 || createuser --superuser $(whoami)"
+# Create the DB user defined in /config/database.yml
+psql postgres -tAc "SELECT 1 FROM pg_roles WHERE rolname='ubuntu'" | grep -q 1 || createuser --superuser basimilchdbuser
 
 # Clean up apt-get files
 sudo apt-get -y update
@@ -76,17 +85,7 @@ echo '# Force rvm to load the right ruby version from the Gemfile on each new tm
 echo 'cd ${PWD}' >> ~/.bashrc
 cd ${PWD}
 
-# Prepare the rails dev environment.
-
-# Install the required ruby version.
-REQUIRED_RUBY_VERSION="$(egrep '^ruby' Gemfile |  sed 's/ruby *.\([0-9.]*\)./\1/')"
-rvm install ruby-${REQUIRED_RUBY_VERSION}
-gem install bundler
-bundle install
-bundle exec rake db:setup
-bundle exec rake test
-
-cat << EOF >> ~/.bashrc
+grep -q 'To launch the application' ~/.bashrc || cat << EOF >> ~/.bashrc
 echo
 echo 'To launch the application:    rails server -b $IP -p $PORT'
 echo 'To launch the console:        rails console'
@@ -98,4 +97,16 @@ echo '                              bundle exec rake db:schema:load RAILS_ENV=te
 echo '                              bundle exec rake db:migrate'
 echo
 EOF
+
+# Prepare the rails dev environment.
+
+# Install the required ruby version.
+REQUIRED_RUBY_VERSION="$(egrep '^ruby' Gemfile |  sed 's/ruby *.\([0-9.]*\)./\1/')"
+rvm install ruby-${REQUIRED_RUBY_VERSION}
+gem install bundler
+bundle install
+bundle exec rake db:setup
+bundle exec rake test
+
+# Source ~/.bashrc to print the help hints.
 source ~/.bashrc
